refactor augmentations
